requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
anthropic>=0.26.0
pymupdf>=1.24.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
python-multipart>=0.0.9
python-dotenv>=1.0.0
Pillow>=10.0.0
